+++ earth2studio/models/px/pangu.py
@@ -26,7 +26,8 @@
 
 import os
 from collections import OrderedDict
-from typing import Generator, Iterator, TypeVar
+from collections.abc import Generator, Iterator
+from typing import TypeVar
 
 import numpy as np
 
@@ -132,7 +133,7 @@ class PanguBase(torch.nn.Module, AutoModelMixin, PrognosticMixin):
 
         self._input_coords = OrderedDict(
             {
-                "batch": np.empty(1),
+                "batch": np.empty(0),
                 "lead_time": np.array([np.timedelta64(0, "h")]),
                 "variable": np.array(VARIABLES),
                 "lat": np.linspace(90, -90, 721, endpoint=True),
@@ -142,7 +143,7 @@ class PanguBase(torch.nn.Module, AutoModelMixin, PrognosticMixin):
 
         self._output_coords = OrderedDict(
             {
-                "batch": np.empty(1),
+                "batch": np.empty(0),
                 "lead_time": np.array([np.timedelta64(6, "h")]),
                 "variable": np.array(VARIABLES),
                 "lat": np.linspace(90, -90, 721, endpoint=True),
@@ -223,8 +224,8 @@ class PanguBase(torch.nn.Module, AutoModelMixin, PrognosticMixin):
         """
         if ort is None:
             raise ImportError(
-                "onnxruntime (onnxruntime-gpu) is required for Pangu. "
-                + "Install the [pangu] optional dependencies"
+                "onnxruntime (onnxruntime-gpu) is required for Pangu. See model install notes for details.\n"
+                + "https://nvidia.github.io/earth2studio/userguide/about/install.html#model-dependencies"
             )
         options = ort.SessionOptions()
         options.enable_cpu_mem_arena = False