+++ test/io/test_kv.py
@@ -15,7 +15,6 @@
 # limitations under the License.
 
 from collections import OrderedDict
-from typing import List
 
 import numpy as np
 import pytest
@@ -23,7 +22,7 @@ import torch
 import xarray as xr
 
 from earth2studio.io import KVBackend
-from earth2studio.utils.coords import extract_coords
+from earth2studio.utils.coords import split_coords
 
 
 @pytest.mark.parametrize(
@@ -38,7 +37,7 @@ from earth2studio.utils.coords import extract_coords
     [["t2m"], ["t2m", "tcwv"]],
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
-def test_kv_fields(time: List[np.datetime64], variable: List[str], device: str) -> None:
+def test_kv_fields(time: list[np.datetime64], variable: list[str], device: str) -> None:
 
     total_coords = OrderedDict(
         {
@@ -121,7 +120,7 @@ def test_kv_fields(time: List[np.datetime64], variable: List[str], device: str) 
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
 def test_kv_variable(
-    time: List[np.datetime64], variable: List[str], device: str
+    time: list[np.datetime64], variable: list[str], device: str
 ) -> None:
 
     total_coords = OrderedDict(
@@ -160,7 +159,7 @@ def test_kv_variable(
     )
     partial_data = torch.randn((1, 1, 180, 180), device=device)
 
-    z.write(*extract_coords(partial_data, partial_coords, "variable"))
+    z.write(*split_coords(partial_data, partial_coords, "variable"))
     assert torch.allclose(z[variable[0]][0, :, :180], partial_data)
 
     # test to xarray
@@ -181,7 +180,7 @@ def test_kv_variable(
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
 def test_kv_exceptions(
-    time: List[np.datetime64], variable: List[str], device: str
+    time: list[np.datetime64], variable: list[str], device: str
 ) -> None:
 
     total_coords = OrderedDict(