+++ test/io/test_netcdf4.py
@@ -17,7 +17,6 @@
 import os
 import tempfile
 from collections import OrderedDict
-from typing import List
 
 import netCDF4
 import numpy as np
@@ -25,7 +24,7 @@ import pytest
 import torch
 
 from earth2studio.io import NetCDF4Backend
-from earth2studio.utils.coords import extract_coords
+from earth2studio.utils.coords import split_coords
 
 
 @pytest.mark.parametrize(
@@ -51,9 +50,9 @@ from earth2studio.utils.coords import extract_coords
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
 def test_netcdf4_fields(
-    time: List[np.datetime64],
-    lead_time: List[np.datetime64],
-    variable: List[str],
+    time: list[np.datetime64],
+    lead_time: list[np.datetime64],
+    variable: list[str],
     device: str,
 ) -> None:
 
@@ -183,7 +182,7 @@ def test_netcdf4_fields(
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
 def test_netcdf4_variable(
-    time: List[np.datetime64], variable: List[str], device: str
+    time: list[np.datetime64], variable: list[str], device: str
 ) -> None:
 
     total_coords = OrderedDict(
@@ -225,7 +224,7 @@ def test_netcdf4_variable(
         }
     )
     partial_data = torch.randn((1, 1, 180, 180), device=device)
-    nc.write(*extract_coords(partial_data, partial_coords, "variable"))
+    nc.write(*split_coords(partial_data, partial_coords, "variable"))
     assert np.allclose(nc[variable[0]][0, :, :180], partial_data.to("cpu").numpy())
     nc.close()
 
@@ -255,7 +254,7 @@ def test_netcdf4_variable(
             }
         )
         partial_data = torch.randn((1, 1, 180, 180), device=device)
-        nc.write(*extract_coords(partial_data, partial_coords, "variable"))
+        nc.write(*split_coords(partial_data, partial_coords, "variable"))
         assert np.allclose(nc[variable[0]][0, :, :180], partial_data.to("cpu").numpy())
         nc.close()
 
@@ -273,7 +272,7 @@ def test_netcdf4_variable(
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
 def test_netcdf4_exceptions(
-    time: List[np.datetime64], variable: List[str], device: str
+    time: list[np.datetime64], variable: list[str], device: str
 ) -> None:
 
     total_coords = OrderedDict(