+++ test/io/test_zarr.py
@@ -17,7 +17,6 @@
 import os
 import tempfile
 from collections import OrderedDict
-from typing import List
 
 import numpy as np
 import pytest
@@ -25,7 +24,7 @@ import torch
 import zarr
 
 from earth2studio.io import ZarrBackend
-from earth2studio.utils.coords import extract_coords
+from earth2studio.utils.coords import split_coords
 
 
 @pytest.mark.parametrize(
@@ -41,7 +40,7 @@ from earth2studio.utils.coords import extract_coords
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
 def test_zarr_field(
-    time: List[np.datetime64], variable: List[str], device: str
+    time: list[np.datetime64], variable: list[str], device: str
 ) -> None:
 
     total_coords = OrderedDict(
@@ -195,7 +194,7 @@ def test_zarr_field(
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
 def test_zarr_variable(
-    time: List[np.datetime64], variable: List[str], device: str
+    time: list[np.datetime64], variable: list[str], device: str
 ) -> None:
 
     total_coords = OrderedDict(
@@ -238,7 +237,7 @@ def test_zarr_variable(
         }
     )
     partial_data = torch.randn((1, 1, 180, 180), device=device)
-    z.write(*extract_coords(partial_data, partial_coords, "variable"))
+    z.write(*split_coords(partial_data, partial_coords, "variable"))
     assert np.allclose(z[variable[0]][0, :, :180], partial_data.to("cpu").numpy())
 
     # Test Directory Store
@@ -271,7 +270,7 @@ def test_zarr_variable(
             }
         )
         partial_data = torch.randn((1, 1, 180, 180), device=device)
-        z.write(*extract_coords(partial_data, partial_coords, "variable"))
+        z.write(*split_coords(partial_data, partial_coords, "variable"))
         assert np.allclose(z[variable[0]][0, :, :180], partial_data.to("cpu").numpy())
 
 
@@ -288,7 +287,7 @@ def test_zarr_variable(
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
 def test_zarr_exceptions(
-    time: List[np.datetime64], variable: List[str], device: str
+    time: list[np.datetime64], variable: list[str], device: str
 ) -> None:
 
     total_coords = OrderedDict(