+++ CHANGELOG.md
@@ -10,7 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Built in diagnostic workflow
+- Basic diagnostic example
+- Batch dimension userguide
+- Parallel inference example
+
 ### Changed
+
+- Changed utility function `extract_coords` to `split_coords`
+- Batched coordinate arrays now use `np.empty(0)` instead of `np.empty(1)`
+- Improving user guide layout and developer documentation
 
 ### Deprecated
 
@@ -18,9 +27,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Enable version switch in documentation site
+- Longitude coordinates of precip and climatenet diagnostic models
+
 ### Security
 
 ### Dependencies
+
+- Bump Modulus required version to 0.6.0
+- PyUpgrade pre-commit hook for Python 3.10
 
 ## [0.1.0] - 2024-04-22